+++ src/test/suite/commands/generateTestFilesCommand.test.ts
@@ -3,7 +3,7 @@ import { describe, it, beforeEach } from 'mocha';
 import { generateTestFilesCommand } from '../../../commands/generateTestFilesCommand';
 import { ConfirmationMode } from '../../../data/configurations/confirmationMode';
 import { SwiftPackageManifest, TargetType } from '../../../data/swiftPackage';
-import { fileUri, fileUris, FullTestFixture, makeExpectedTestFileContentString, stubPackage } from '../fullTestFixture';
+import { fileUri, fileUris, FullTestFixture, makeExpectedTestFileContentString, stubPackage, ShowFileArguments } from '../fullTestFixture';
 
 suite('generateTestFilesCommand Test Suite', () => {
     describe('generateTestFilesCommand', () => {
@@ -32,6 +32,27 @@ suite('generateTestFilesCommand Test Suite', () => {
                     fileContents: makeExpectedTestFileContentString("Target", "BTests")
                 },
             ]);
+        });
+
+        it('should open the first newly created file on the editor', async () => {
+            const files = fileUris(
+                "/home/Sources/Target/A.swift",
+                "/home/Sources/Target/B.swift",
+            );
+            const pkg = stubPackage();
+            const fixture = new FullTestFixture([
+                "/home/Package.swift",
+                "/home/Sources/Target/A.swift",
+                "/home/Sources/Target/B.swift",
+                "/home/Tests/TargetTests/",
+            ], undefined, pkg);
+            fixture.context.setCreateVirtualFilesInWorkspaceEdits(true);
+
+            await generateTestFilesCommand(files, ConfirmationMode.always, fixture.context);
+
+            fixture.assertShownFiles({
+                fileUri: "/home/Tests/TargetTests/ATests.swift"
+            });
         });
 
         it('must de-duplicate file uris before generating test files', async () => {
@@ -129,6 +150,61 @@ suite('generateTestFilesCommand Test Suite', () => {
                     fileContents: makeExpectedTestFileContentString("Target", "ATests")
                 },
             ]);
+        });
+
+        it('must not create test files that already exist', async () => {
+            const files = fileUris(
+                "/home/Sources/Target/A.swift",
+            );
+            const pkg = stubPackage();
+            const fixture = new FullTestFixture([
+                "/home/Package.swift",
+                "/home/Sources/Target/A.swift",
+                "/home/Tests/TargetTests/",
+                "/home/Tests/TargetTests/ATests.swift",
+            ], undefined, pkg);
+
+            await generateTestFilesCommand(files, ConfirmationMode.always, fixture.context);
+
+            fixture.assertNoModifyingWorkspaceEdits();
+        });
+
+        it('should open an existing test file if it already exists', async () => {
+            const files = fileUris(
+                "/home/Sources/Target/A.swift",
+            );
+            const pkg = stubPackage();
+            const fixture = new FullTestFixture([
+                "/home/Package.swift",
+                "/home/Sources/Target/A.swift",
+                "/home/Tests/TargetTests/",
+                "/home/Tests/TargetTests/ATests.swift",
+            ], undefined, pkg);
+
+            await generateTestFilesCommand(files, ConfirmationMode.always, fixture.context);
+
+            fixture.assertShownFiles(
+                {
+                    fileUri: fileUri("/home/Tests/TargetTests/ATests.swift")
+                },
+            );
+        });
+
+        it('should not issue an open command for a test file if the user cancels the edit and no file is created', async () => {
+            const files = fileUris(
+                "/home/Sources/Target/A.swift",
+            );
+            const pkg = stubPackage();
+            const fixture = new FullTestFixture([
+                "/home/Package.swift",
+                "/home/Sources/Target/A.swift",
+                "/home/Tests/TargetTests/",
+            ], undefined, pkg);
+            fixture.context.setCreateVirtualFilesInWorkspaceEdits(false);
+
+            await generateTestFilesCommand(files, ConfirmationMode.always, fixture.context);
+
+            fixture.assertShownFiles();
         });
 
         it('should do nothing for files not in recognized sources folder', async () => {