+++ src/test/suite/commands/gotoTestFileCommand.test.ts
@@ -22,7 +22,10 @@ suite('gotoTestFileCommand Test Suite', () => {
             await gotoTestFileCommand(file, fixture.context);
 
             fixture.assertShownFiles(
-                ["/home/Tests/TargetTests/ATests.swift", { viewColumn: vscode.ViewColumn.Active }],
+                {
+                    fileUri: "/home/Tests/TargetTests/ATests.swift",
+                    options: { viewColumn: vscode.ViewColumn.Active }
+                },
             );
         });
 
@@ -88,7 +91,7 @@ suite('gotoTestFileCommand Test Suite', () => {
                         fileContents: makeExpectedTestFileContentString("Target", "ATests")
                     },
                 ])
-                .assertShownFiles(['/home/Tests/TargetTests/ATests.swift']);
+                .assertShownFiles({fileUri: '/home/Tests/TargetTests/ATests.swift'});
         });
 
         it('should not create a file when none is found, if the user chooses not to.', async () => {
@@ -138,7 +141,10 @@ suite('gotoTestFileCommand Test Suite', () => {
                 await gotoTestFileCommand(file, fixture.context);
 
                 fixture.assertShownFiles(
-                    ["/home/Tests/TargetTests/ATestFile.swift", { viewColumn: vscode.ViewColumn.Active }],
+                    {
+                        fileUri: "/home/Tests/TargetTests/ATestFile.swift",
+                        options: { viewColumn: vscode.ViewColumn.Active }
+                    },
                 );
             });
 
@@ -197,9 +203,18 @@ suite('gotoTestFileCommand Test Suite', () => {
                 ), fixture.context);
 
                 fixture.assertShownFiles(
-                    ["/home/Tests/TargetTests/ATestFile.swift", { viewColumn: vscode.ViewColumn.Active }],
-                    ["/home/Tests/TargetTests/BSpec.swift", { viewColumn: vscode.ViewColumn.Active }],
-                    ["/home/Tests/TargetTests/CTests.swift", { viewColumn: vscode.ViewColumn.Active }],
+                    {
+                        fileUri: "/home/Tests/TargetTests/ATestFile.swift",
+                        options: { viewColumn: vscode.ViewColumn.Active }
+                    },
+                    {
+                        fileUri: "/home/Tests/TargetTests/BSpec.swift",
+                        options: { viewColumn: vscode.ViewColumn.Active }
+                    },
+                    {
+                        fileUri: "/home/Tests/TargetTests/CTests.swift",
+                        options: { viewColumn: vscode.ViewColumn.Active }
+                    },
                 );
             });
 
@@ -233,7 +248,7 @@ suite('gotoTestFileCommand Test Suite', () => {
                             fileContents: makeExpectedTestFileContentString("Target", "ATests")
                         },
                     ])
-                    .assertShownFiles(['/home/Tests/TargetTests/ATests.swift']);
+                    .assertShownFiles({fileUri: '/home/Tests/TargetTests/ATests.swift'});
             });
 
             it('should not create a file when none is found, if the user chooses not to.', async () => {
@@ -275,7 +290,10 @@ suite('gotoTestFileCommand Test Suite', () => {
 
                 fixture
                     .assertShownFiles(
-                        ["/home/Tests/TargetTests/ATestFile.swift", { viewColumn: vscode.ViewColumn.Active }],
+                        {
+                            fileUri: "/home/Tests/TargetTests/ATestFile.swift",
+                            options: { viewColumn: vscode.ViewColumn.Active }
+                        },
                     )
                     .assertShownError();
             });